Human_Simulation/PythonCode/requirements.txt
dash==2.7.0
dash-bootstrap-components==1.2.1
Pillow==9.0.0
pandas==1.3.4
numpy==1.20.3
plotly==5.9.0
colorlover==0.3.0
waitress==2.1.2